refactor: dataResponse 제거 및 EstatesResponse추가
