+++ src/main/java/com/cotato/networking1/estate/service/EstateService.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.cotato.networking1.common.dto.DataResponse;
+import com.cotato.networking1.estate.dto.response.EstatesResponse;
 import com.cotato.networking1.estate.domain.Estate;
-import com.cotato.networking1.estate.dto.EstateResponse;
+import com.cotato.networking1.estate.dto.response.EstateResponse;
 import com.cotato.networking1.estate.repository.EstateRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -21,12 +21,12 @@ public class EstateService {
 		return estateRepository.findAllByZipCode(zipCode);
 	}
 
-	public DataResponse<List<EstateResponse>> makeEstateResponses(int zipCode) {
+	public EstatesResponse makeEstateResponses(int zipCode) {
 		List<EstateResponse> responses = findEstatesByZipCode(zipCode).stream()
 			.map(EstateResponse::create)
 			.toList();
 
-		return DataResponse.ok(responses);
+		return EstatesResponse.create(responses);
 	}
